video_core: track unified wrapper hazards per bucket instead of whole-buffer
The unified wrapper shares one (access, stage) state across 4+GB, so the
write-involved no-dedup rule made barrier emissions scale with bind count:
nearly every descriptor bind emitted an eAllCommands whole-tail barrier and
split the current render pass. Replace it with per-4MiB last-writer buckets:
reads barrier once per producer->consumer edge (publish wide, retire the
writer), writes order against the prior writer and readers, and untouched
buckets emit nothing.

Vertex, index, and indirect fetches of unified-served ranges now always
consult the tracker; unified writes never mark the memory tracker, so the
old IsRegionGpuModified gates never fired for wrapper-written data and
those fetches ran with no barrier at all once the storm stopped.

Fill publishes itself with explicit pre/post barriers and is recorded as
self-synced. Tile manager offsets widen to u64 for wrapper physical
offsets, and the tiler gains its missing post-dispatch availability
barrier. native_uma bit 3 (8) reverts to the previous whole-wrapper rule
for A/B comparison.

+++ src/video_core/buffer_cache/buffer.h
@@ -131,7 +131,7 @@ public:
 
     std::optional<vk::BufferMemoryBarrier2> GetBarrier(vk::AccessFlags2 dst_acess_mask,
                                                        vk::PipelineStageFlagBits2 dst_stage,
-                                                       u32 offset = 0) {
+                                                       u64 offset = 0) {
         if (dst_acess_mask == access_mask && stage == dst_stage) {
             // The unified wrapper aliases every guest range under one state;
             // equal-state deduplication would suppress the barrier between a
